third_party/WebKit/LayoutTests/fast/forms/select/option-innertext.html

<!DOCTYPE html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<option>Text</option>
<script>
test(() => {
  assert_equals(document.querySelector('option').innerText, 'Text');
}, 'option.innerText should not return an empty string if it has non-empty text child.');
</script>
